pyproject.toml
[project]
name = "ies-optimization"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8,<3.9"
dependencies = [
    "geatpy==2.7.0",
    "gurobipy==11.0.3",
    "matplotlib==3.3.4",
    "numpy==1.24.4",
    "oemof-network==0.4.0",
    "oemof-solph==0.4.4",
    "openpyxl>=3.1.5",
    "pandas==1.2.4",
    "pymupdf>=1.24.11",
    "scipy>=1.10.1",
    "seaborn>=0.13.0",
    "xlrd>=2.0.2",
]
